Improved release process
Included -r option in buildpackage.sh to prepare a source release for a
given Ubuntu distro, ready to be uploaded to a Ubuntu PPA. Included -rn
option to remove Vala as build dependency in build-package.sh and do
everything that -r does.

Modified configure.ac to avoid checking for Vala compiler if the C
sources have already been generated.